Move font.ts to assets.
