+++ docs/api/classes/CartesianCoordinateSystem.html
@@ -62,12 +62,12 @@
 <h3 class="tsd-anchor-link"><code class="tsd-tag ts-flagReadonly">Readonly</code> <span>xscale</span><a href="#xscale" aria-label="Permalink" class="tsd-anchor-icon"><svg class="icon icon-tabler icon-tabler-link" viewBox="0 0 24 24" stroke-width="2" stroke="currentColor" fill="none" stroke-linecap="round" stroke-linejoin="round"><use href="#icon-anchor-a"></use><use href="#icon-anchor-b"></use><use href="#icon-anchor-c"></use></svg></a></h3>
 <div class="tsd-signature">xscale<span class="tsd-signature-symbol">:</span> <a href="LinearScale.html" class="tsd-signature-type" data-tsd-kind="Class">LinearScale</a><span class="tsd-signature-symbol"> | </span><a href="LogScale.html" class="tsd-signature-type" data-tsd-kind="Class">LogScale</a></div><aside class="tsd-sources">
 <ul>
-<li>Defined in <a href="https://github.com/wwwtyro/candygraph/blob/0dd03ce/src/coordinates/cartesian.ts#L27">coordinates/cartesian.ts:27</a></li></ul></aside></section>
+<li>Defined in <a href="https://github.com/wwwtyro/candygraph/blob/f289c4d/src/coordinates/cartesian.ts#L27">coordinates/cartesian.ts:27</a></li></ul></aside></section>
 <section class="tsd-panel tsd-member tsd-kind-property tsd-parent-kind-class"><a id="yscale" class="tsd-anchor"></a>
 <h3 class="tsd-anchor-link"><code class="tsd-tag ts-flagReadonly">Readonly</code> <span>yscale</span><a href="#yscale" aria-label="Permalink" class="tsd-anchor-icon"><svg class="icon icon-tabler icon-tabler-link" viewBox="0 0 24 24" stroke-width="2" stroke="currentColor" fill="none" stroke-linecap="round" stroke-linejoin="round"><use href="#icon-anchor-a"></use><use href="#icon-anchor-b"></use><use href="#icon-anchor-c"></use></svg></a></h3>
 <div class="tsd-signature">yscale<span class="tsd-signature-symbol">:</span> <a href="LinearScale.html" class="tsd-signature-type" data-tsd-kind="Class">LinearScale</a><span class="tsd-signature-symbol"> | </span><a href="LogScale.html" class="tsd-signature-type" data-tsd-kind="Class">LogScale</a></div><aside class="tsd-sources">
 <ul>
-<li>Defined in <a href="https://github.com/wwwtyro/candygraph/blob/0dd03ce/src/coordinates/cartesian.ts#L28">coordinates/cartesian.ts:28</a></li></ul></aside></section></section>
+<li>Defined in <a href="https://github.com/wwwtyro/candygraph/blob/f289c4d/src/coordinates/cartesian.ts#L28">coordinates/cartesian.ts:28</a></li></ul></aside></section></section>
 <section class="tsd-panel-group tsd-member-group">
 <h2>Methods</h2>
 <section class="tsd-panel tsd-member tsd-kind-method tsd-parent-kind-class"><a id="toDomain" class="tsd-anchor"></a>
