+++ docs/api/classes/HLines.html
@@ -72,22 +72,22 @@ inconsistent appearance.</p>
 <h3 class="tsd-anchor-link"><code class="tsd-tag ts-flagReadonly">Readonly</code> <span>colors</span><a href="#colors" aria-label="Permalink" class="tsd-anchor-icon"><svg class="icon icon-tabler icon-tabler-link" viewBox="0 0 24 24" stroke-width="2" stroke="currentColor" fill="none" stroke-linecap="round" stroke-linejoin="round"><use href="#icon-anchor-a"></use><use href="#icon-anchor-b"></use><use href="#icon-anchor-c"></use></svg></a></h3>
 <div class="tsd-signature">colors<span class="tsd-signature-symbol">:</span> <a href="Dataset.html" class="tsd-signature-type" data-tsd-kind="Class">Dataset</a></div><aside class="tsd-sources">
 <ul>
-<li>Defined in <a href="https://github.com/wwwtyro/candygraph/blob/0dd03ce/src/primitives/hlines.ts#L41">primitives/hlines.ts:41</a></li></ul></aside></section>
+<li>Defined in <a href="https://github.com/wwwtyro/candygraph/blob/f289c4d/src/primitives/hlines.ts#L41">primitives/hlines.ts:41</a></li></ul></aside></section>
 <section class="tsd-panel tsd-member tsd-kind-property tsd-parent-kind-class"><a id="lines" class="tsd-anchor"></a>
 <h3 class="tsd-anchor-link"><code class="tsd-tag ts-flagReadonly">Readonly</code> <span>lines</span><a href="#lines" aria-label="Permalink" class="tsd-anchor-icon"><svg class="icon icon-tabler icon-tabler-link" viewBox="0 0 24 24" stroke-width="2" stroke="currentColor" fill="none" stroke-linecap="round" stroke-linejoin="round"><use href="#icon-anchor-a"></use><use href="#icon-anchor-b"></use><use href="#icon-anchor-c"></use></svg></a></h3>
 <div class="tsd-signature">lines<span class="tsd-signature-symbol">:</span> <a href="Dataset.html" class="tsd-signature-type" data-tsd-kind="Class">Dataset</a></div><aside class="tsd-sources">
 <ul>
-<li>Defined in <a href="https://github.com/wwwtyro/candygraph/blob/0dd03ce/src/primitives/hlines.ts#L39">primitives/hlines.ts:39</a></li></ul></aside></section>
+<li>Defined in <a href="https://github.com/wwwtyro/candygraph/blob/f289c4d/src/primitives/hlines.ts#L39">primitives/hlines.ts:39</a></li></ul></aside></section>
 <section class="tsd-panel tsd-member tsd-kind-property tsd-parent-kind-class tsd-is-private"><a id="segmentGeometry" class="tsd-anchor"></a>
 <h3 class="tsd-anchor-link"><code class="tsd-tag ts-flagPrivate">Private</code> <span>segment<wbr/>Geometry</span><a href="#segmentGeometry" aria-label="Permalink" class="tsd-anchor-icon"><svg class="icon icon-tabler icon-tabler-link" viewBox="0 0 24 24" stroke-width="2" stroke="currentColor" fill="none" stroke-linecap="round" stroke-linejoin="round"><use href="#icon-anchor-a"></use><use href="#icon-anchor-b"></use><use href="#icon-anchor-c"></use></svg></a></h3>
 <div class="tsd-signature">segment<wbr/>Geometry<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">Buffer</span></div><aside class="tsd-sources">
 <ul>
-<li>Defined in <a href="https://github.com/wwwtyro/candygraph/blob/0dd03ce/src/primitives/hlines.ts#L42">primitives/hlines.ts:42</a></li></ul></aside></section>
+<li>Defined in <a href="https://github.com/wwwtyro/candygraph/blob/f289c4d/src/primitives/hlines.ts#L42">primitives/hlines.ts:42</a></li></ul></aside></section>
 <section class="tsd-panel tsd-member tsd-kind-property tsd-parent-kind-class"><a id="widths" class="tsd-anchor"></a>
 <h3 class="tsd-anchor-link"><code class="tsd-tag ts-flagReadonly">Readonly</code> <span>widths</span><a href="#widths" aria-label="Permalink" class="tsd-anchor-icon"><svg class="icon icon-tabler icon-tabler-link" viewBox="0 0 24 24" stroke-width="2" stroke="currentColor" fill="none" stroke-linecap="round" stroke-linejoin="round"><use href="#icon-anchor-a"></use><use href="#icon-anchor-b"></use><use href="#icon-anchor-c"></use></svg></a></h3>
 <div class="tsd-signature">widths<span class="tsd-signature-symbol">:</span> <a href="Dataset.html" class="tsd-signature-type" data-tsd-kind="Class">Dataset</a></div><aside class="tsd-sources">
 <ul>
-<li>Defined in <a href="https://github.com/wwwtyro/candygraph/blob/0dd03ce/src/primitives/hlines.ts#L40">primitives/hlines.ts:40</a></li></ul></aside></section></section>
+<li>Defined in <a href="https://github.com/wwwtyro/candygraph/blob/f289c4d/src/primitives/hlines.ts#L40">primitives/hlines.ts:40</a></li></ul></aside></section></section>
 <section class="tsd-panel-group tsd-member-group">
 <h2>Methods</h2>
 <section class="tsd-panel tsd-member tsd-kind-method tsd-parent-kind-class"><a id="dispose" class="tsd-anchor"></a>
