+++ docs/api/classes/PolarCoordinateSystem.html
@@ -68,22 +68,22 @@
 <h3 class="tsd-anchor-link"><code class="tsd-tag ts-flagReadonly">Readonly</code> <span>angular<wbr/>Scale</span><a href="#angularScale" aria-label="Permalink" class="tsd-anchor-icon"><svg class="icon icon-tabler icon-tabler-link" viewBox="0 0 24 24" stroke-width="2" stroke="currentColor" fill="none" stroke-linecap="round" stroke-linejoin="round"><use href="#icon-anchor-a"></use><use href="#icon-anchor-b"></use><use href="#icon-anchor-c"></use></svg></a></h3>
 <div class="tsd-signature">angular<wbr/>Scale<span class="tsd-signature-symbol">:</span> <a href="LinearScale.html" class="tsd-signature-type" data-tsd-kind="Class">LinearScale</a><span class="tsd-signature-symbol"> | </span><a href="LogScale.html" class="tsd-signature-type" data-tsd-kind="Class">LogScale</a></div><aside class="tsd-sources">
 <ul>
-<li>Defined in <a href="https://github.com/wwwtyro/candygraph/blob/0dd03ce/src/coordinates/polar.ts#L33">coordinates/polar.ts:33</a></li></ul></aside></section>
+<li>Defined in <a href="https://github.com/wwwtyro/candygraph/blob/f289c4d/src/coordinates/polar.ts#L33">coordinates/polar.ts:33</a></li></ul></aside></section>
 <section class="tsd-panel tsd-member tsd-kind-property tsd-parent-kind-class"><a id="radialScale" class="tsd-anchor"></a>
 <h3 class="tsd-anchor-link"><code class="tsd-tag ts-flagReadonly">Readonly</code> <span>radial<wbr/>Scale</span><a href="#radialScale" aria-label="Permalink" class="tsd-anchor-icon"><svg class="icon icon-tabler icon-tabler-link" viewBox="0 0 24 24" stroke-width="2" stroke="currentColor" fill="none" stroke-linecap="round" stroke-linejoin="round"><use href="#icon-anchor-a"></use><use href="#icon-anchor-b"></use><use href="#icon-anchor-c"></use></svg></a></h3>
 <div class="tsd-signature">radial<wbr/>Scale<span class="tsd-signature-symbol">:</span> <a href="LinearScale.html" class="tsd-signature-type" data-tsd-kind="Class">LinearScale</a><span class="tsd-signature-symbol"> | </span><a href="LogScale.html" class="tsd-signature-type" data-tsd-kind="Class">LogScale</a></div><aside class="tsd-sources">
 <ul>
-<li>Defined in <a href="https://github.com/wwwtyro/candygraph/blob/0dd03ce/src/coordinates/polar.ts#L32">coordinates/polar.ts:32</a></li></ul></aside></section>
+<li>Defined in <a href="https://github.com/wwwtyro/candygraph/blob/f289c4d/src/coordinates/polar.ts#L32">coordinates/polar.ts:32</a></li></ul></aside></section>
 <section class="tsd-panel tsd-member tsd-kind-property tsd-parent-kind-class"><a id="xScale" class="tsd-anchor"></a>
 <h3 class="tsd-anchor-link"><code class="tsd-tag ts-flagReadonly">Readonly</code> <span>x<wbr/>Scale</span><a href="#xScale" aria-label="Permalink" class="tsd-anchor-icon"><svg class="icon icon-tabler icon-tabler-link" viewBox="0 0 24 24" stroke-width="2" stroke="currentColor" fill="none" stroke-linecap="round" stroke-linejoin="round"><use href="#icon-anchor-a"></use><use href="#icon-anchor-b"></use><use href="#icon-anchor-c"></use></svg></a></h3>
 <div class="tsd-signature">x<wbr/>Scale<span class="tsd-signature-symbol">:</span> <a href="LinearScale.html" class="tsd-signature-type" data-tsd-kind="Class">LinearScale</a><span class="tsd-signature-symbol"> | </span><a href="LogScale.html" class="tsd-signature-type" data-tsd-kind="Class">LogScale</a></div><aside class="tsd-sources">
 <ul>
-<li>Defined in <a href="https://github.com/wwwtyro/candygraph/blob/0dd03ce/src/coordinates/polar.ts#L34">coordinates/polar.ts:34</a></li></ul></aside></section>
+<li>Defined in <a href="https://github.com/wwwtyro/candygraph/blob/f289c4d/src/coordinates/polar.ts#L34">coordinates/polar.ts:34</a></li></ul></aside></section>
 <section class="tsd-panel tsd-member tsd-kind-property tsd-parent-kind-class"><a id="yScale" class="tsd-anchor"></a>
 <h3 class="tsd-anchor-link"><code class="tsd-tag ts-flagReadonly">Readonly</code> <span>y<wbr/>Scale</span><a href="#yScale" aria-label="Permalink" class="tsd-anchor-icon"><svg class="icon icon-tabler icon-tabler-link" viewBox="0 0 24 24" stroke-width="2" stroke="currentColor" fill="none" stroke-linecap="round" stroke-linejoin="round"><use href="#icon-anchor-a"></use><use href="#icon-anchor-b"></use><use href="#icon-anchor-c"></use></svg></a></h3>
 <div class="tsd-signature">y<wbr/>Scale<span class="tsd-signature-symbol">:</span> <a href="LinearScale.html" class="tsd-signature-type" data-tsd-kind="Class">LinearScale</a><span class="tsd-signature-symbol"> | </span><a href="LogScale.html" class="tsd-signature-type" data-tsd-kind="Class">LogScale</a></div><aside class="tsd-sources">
 <ul>
-<li>Defined in <a href="https://github.com/wwwtyro/candygraph/blob/0dd03ce/src/coordinates/polar.ts#L35">coordinates/polar.ts:35</a></li></ul></aside></section></section>
+<li>Defined in <a href="https://github.com/wwwtyro/candygraph/blob/f289c4d/src/coordinates/polar.ts#L35">coordinates/polar.ts:35</a></li></ul></aside></section></section>
 <section class="tsd-panel-group tsd-member-group">
 <h2>Methods</h2>
 <section class="tsd-panel tsd-member tsd-kind-method tsd-parent-kind-class"><a id="toDomain" class="tsd-anchor"></a>
