+++ docs/api/interfaces/HLinesOptions.html
@@ -33,12 +33,12 @@
 <h3 class="tsd-anchor-link"><code class="tsd-tag ts-flagOptional">Optional</code> <span>colors</span><a href="#colors" aria-label="Permalink" class="tsd-anchor-icon"><svg class="icon icon-tabler icon-tabler-link" viewBox="0 0 24 24" stroke-width="2" stroke="currentColor" fill="none" stroke-linecap="round" stroke-linejoin="round"><path stroke="none" d="M0 0h24v24H0z" fill="none" id="icon-anchor-a"></path><path d="M10 14a3.5 3.5 0 0 0 5 0l4 -4a3.5 3.5 0 0 0 -5 -5l-.5 .5" id="icon-anchor-b"></path><path d="M14 10a3.5 3.5 0 0 0 -5 0l-4 4a3.5 3.5 0 0 0 5 5l.5 -.5" id="icon-anchor-c"></path></svg></a></h3>
 <div class="tsd-signature">colors<span class="tsd-signature-symbol">?:</span> <a href="../types/NumberArray.html" class="tsd-signature-type" data-tsd-kind="Type alias">NumberArray</a><span class="tsd-signature-symbol"> | </span><a href="../classes/Dataset.html" class="tsd-signature-type" data-tsd-kind="Class">Dataset</a></div><aside class="tsd-sources">
 <ul>
-<li>Defined in <a href="https://github.com/wwwtyro/candygraph/blob/0dd03ce/src/primitives/hlines.ts#L8">primitives/hlines.ts:8</a></li></ul></aside></section>
+<li>Defined in <a href="https://github.com/wwwtyro/candygraph/blob/f289c4d/src/primitives/hlines.ts#L8">primitives/hlines.ts:8</a></li></ul></aside></section>
 <section class="tsd-panel tsd-member tsd-kind-property tsd-parent-kind-interface"><a id="widths" class="tsd-anchor"></a>
 <h3 class="tsd-anchor-link"><code class="tsd-tag ts-flagOptional">Optional</code> <span>widths</span><a href="#widths" aria-label="Permalink" class="tsd-anchor-icon"><svg class="icon icon-tabler icon-tabler-link" viewBox="0 0 24 24" stroke-width="2" stroke="currentColor" fill="none" stroke-linecap="round" stroke-linejoin="round"><use href="#icon-anchor-a"></use><use href="#icon-anchor-b"></use><use href="#icon-anchor-c"></use></svg></a></h3>
 <div class="tsd-signature">widths<span class="tsd-signature-symbol">?:</span> <span class="tsd-signature-type">number</span><span class="tsd-signature-symbol"> | </span><a href="../types/NumberArray.html" class="tsd-signature-type" data-tsd-kind="Type alias">NumberArray</a><span class="tsd-signature-symbol"> | </span><a href="../classes/Dataset.html" class="tsd-signature-type" data-tsd-kind="Class">Dataset</a></div><aside class="tsd-sources">
 <ul>
-<li>Defined in <a href="https://github.com/wwwtyro/candygraph/blob/0dd03ce/src/primitives/hlines.ts#L7">primitives/hlines.ts:7</a></li></ul></aside></section></section></div>
+<li>Defined in <a href="https://github.com/wwwtyro/candygraph/blob/f289c4d/src/primitives/hlines.ts#L7">primitives/hlines.ts:7</a></li></ul></aside></section></section></div>
 <div class="col-4 col-menu menu-sticky-wrap menu-highlight">
 <div class="tsd-navigation settings">
 <details class="tsd-index-accordion"><summary class="tsd-accordion-summary">
