+++ docs/api/interfaces/OrthoAxisInfo.html
@@ -35,14 +35,14 @@
 <div class="tsd-comment tsd-typography"><p>The position of the minor ticks of this axis.</p>
 </div><aside class="tsd-sources">
 <ul>
-<li>Defined in <a href="https://github.com/wwwtyro/candygraph/blob/0dd03ce/src/composites/ortho-axis.ts#L36">composites/ortho-axis.ts:36</a></li></ul></aside></section>
+<li>Defined in <a href="https://github.com/wwwtyro/candygraph/blob/f289c4d/src/composites/ortho-axis.ts#L36">composites/ortho-axis.ts:36</a></li></ul></aside></section>
 <section class="tsd-panel tsd-member tsd-kind-property tsd-parent-kind-interface"><a id="ticks" class="tsd-anchor"></a>
 <h3 class="tsd-anchor-link"><span>ticks</span><a href="#ticks" aria-label="Permalink" class="tsd-anchor-icon"><svg class="icon icon-tabler icon-tabler-link" viewBox="0 0 24 24" stroke-width="2" stroke="currentColor" fill="none" stroke-linecap="round" stroke-linejoin="round"><use href="#icon-anchor-a"></use><use href="#icon-anchor-b"></use><use href="#icon-anchor-c"></use></svg></a></h3>
 <div class="tsd-signature">ticks<span class="tsd-signature-symbol">:</span> <a href="../types/NumberArray.html" class="tsd-signature-type" data-tsd-kind="Type alias">NumberArray</a></div>
 <div class="tsd-comment tsd-typography"><p>The position of the major ticks of this axis.</p>
 </div><aside class="tsd-sources">
 <ul>
-<li>Defined in <a href="https://github.com/wwwtyro/candygraph/blob/0dd03ce/src/composites/ortho-axis.ts#L34">composites/ortho-axis.ts:34</a></li></ul></aside></section></section></div>
+<li>Defined in <a href="https://github.com/wwwtyro/candygraph/blob/f289c4d/src/composites/ortho-axis.ts#L34">composites/ortho-axis.ts:34</a></li></ul></aside></section></section></div>
 <div class="col-4 col-menu menu-sticky-wrap menu-highlight">
 <div class="tsd-navigation settings">
 <details class="tsd-index-accordion"><summary class="tsd-accordion-summary">
