+++ docs/api/interfaces/ShapesOptions.html
@@ -37,7 +37,7 @@
 to all the shapes. Default [0, 0, 0, 0.5].</p>
 </div><aside class="tsd-sources">
 <ul>
-<li>Defined in <a href="https://github.com/wwwtyro/candygraph/blob/0dd03ce/src/primitives/shapes.ts#L9">primitives/shapes.ts:9</a></li></ul></aside></section>
+<li>Defined in <a href="https://github.com/wwwtyro/candygraph/blob/f289c4d/src/primitives/shapes.ts#L9">primitives/shapes.ts:9</a></li></ul></aside></section>
 <section class="tsd-panel tsd-member tsd-kind-property tsd-parent-kind-interface"><a id="rotations" class="tsd-anchor"></a>
 <h3 class="tsd-anchor-link"><code class="tsd-tag ts-flagOptional">Optional</code> <span>rotations</span><a href="#rotations" aria-label="Permalink" class="tsd-anchor-icon"><svg class="icon icon-tabler icon-tabler-link" viewBox="0 0 24 24" stroke-width="2" stroke="currentColor" fill="none" stroke-linecap="round" stroke-linejoin="round"><use href="#icon-anchor-a"></use><use href="#icon-anchor-b"></use><use href="#icon-anchor-c"></use></svg></a></h3>
 <div class="tsd-signature">rotations<span class="tsd-signature-symbol">?:</span> <span class="tsd-signature-type">number</span><span class="tsd-signature-symbol"> | </span><a href="../types/NumberArray.html" class="tsd-signature-type" data-tsd-kind="Type alias">NumberArray</a><span class="tsd-signature-symbol"> | </span><a href="../classes/Dataset.html" class="tsd-signature-type" data-tsd-kind="Class">Dataset</a></div>
@@ -45,7 +45,7 @@ to all the shapes. Default [0, 0, 0, 0.5].</p>
 will apply to all the shapes. Default 0.</p>
 </div><aside class="tsd-sources">
 <ul>
-<li>Defined in <a href="https://github.com/wwwtyro/candygraph/blob/0dd03ce/src/primitives/shapes.ts#L15">primitives/shapes.ts:15</a></li></ul></aside></section>
+<li>Defined in <a href="https://github.com/wwwtyro/candygraph/blob/f289c4d/src/primitives/shapes.ts#L15">primitives/shapes.ts:15</a></li></ul></aside></section>
 <section class="tsd-panel tsd-member tsd-kind-property tsd-parent-kind-interface"><a id="scales" class="tsd-anchor"></a>
 <h3 class="tsd-anchor-link"><code class="tsd-tag ts-flagOptional">Optional</code> <span>scales</span><a href="#scales" aria-label="Permalink" class="tsd-anchor-icon"><svg class="icon icon-tabler icon-tabler-link" viewBox="0 0 24 24" stroke-width="2" stroke="currentColor" fill="none" stroke-linecap="round" stroke-linejoin="round"><use href="#icon-anchor-a"></use><use href="#icon-anchor-b"></use><use href="#icon-anchor-c"></use></svg></a></h3>
 <div class="tsd-signature">scales<span class="tsd-signature-symbol">?:</span> <a href="../types/NumberArray.html" class="tsd-signature-type" data-tsd-kind="Type alias">NumberArray</a><span class="tsd-signature-symbol"> | </span><a href="../classes/Dataset.html" class="tsd-signature-type" data-tsd-kind="Class">Dataset</a></div>
@@ -53,7 +53,7 @@ will apply to all the shapes. Default 0.</p>
 to all the shapes. Default [1, 1].</p>
 </div><aside class="tsd-sources">
 <ul>
-<li>Defined in <a href="https://github.com/wwwtyro/candygraph/blob/0dd03ce/src/primitives/shapes.ts#L12">primitives/shapes.ts:12</a></li></ul></aside></section></section></div>
+<li>Defined in <a href="https://github.com/wwwtyro/candygraph/blob/f289c4d/src/primitives/shapes.ts#L12">primitives/shapes.ts:12</a></li></ul></aside></section></section></div>
 <div class="col-4 col-menu menu-sticky-wrap menu-highlight">
 <div class="tsd-navigation settings">
 <details class="tsd-index-accordion"><summary class="tsd-accordion-summary">
