+++ docs/api/interfaces/TransparentLineStripOptions.html
@@ -35,14 +35,14 @@
 <div class="tsd-comment tsd-typography"><p>The color of the line strip. Default [0, 0, 0, 1].</p>
 </div><aside class="tsd-sources">
 <ul>
-<li>Defined in primitives/transparent-line-strip.ts:11</li></ul></aside></section>
+<li>Defined in <a href="https://github.com/wwwtyro/candygraph/blob/f289c4d/src/primitives/transparent-line-strip.ts#L11">primitives/transparent-line-strip.ts:11</a></li></ul></aside></section>
 <section class="tsd-panel tsd-member tsd-kind-property tsd-parent-kind-interface"><a id="width" class="tsd-anchor"></a>
 <h3 class="tsd-anchor-link"><code class="tsd-tag ts-flagOptional">Optional</code> <span>width</span><a href="#width" aria-label="Permalink" class="tsd-anchor-icon"><svg class="icon icon-tabler icon-tabler-link" viewBox="0 0 24 24" stroke-width="2" stroke="currentColor" fill="none" stroke-linecap="round" stroke-linejoin="round"><use href="#icon-anchor-a"></use><use href="#icon-anchor-b"></use><use href="#icon-anchor-c"></use></svg></a></h3>
 <div class="tsd-signature">width<span class="tsd-signature-symbol">?:</span> <span class="tsd-signature-type">number</span></div>
 <div class="tsd-comment tsd-typography"><p>The width of the line strip. Default 1.</p>
 </div><aside class="tsd-sources">
 <ul>
-<li>Defined in primitives/transparent-line-strip.ts:8</li></ul></aside></section></section></div>
+<li>Defined in <a href="https://github.com/wwwtyro/candygraph/blob/f289c4d/src/primitives/transparent-line-strip.ts#L8">primitives/transparent-line-strip.ts:8</a></li></ul></aside></section></section></div>
 <div class="col-4 col-menu menu-sticky-wrap menu-highlight">
 <div class="tsd-navigation settings">
 <details class="tsd-index-accordion"><summary class="tsd-accordion-summary">
