+++ docs/api/interfaces/TrianglesOptions.html
@@ -34,7 +34,7 @@
 <div class="tsd-comment tsd-typography"><p>The color of the triangles. Default [0, 0, 0, 0.5].</p>
 </div><aside class="tsd-sources">
 <ul>
-<li>Defined in <a href="https://github.com/wwwtyro/candygraph/blob/0dd03ce/src/primitives/triangles.ts#L8">primitives/triangles.ts:8</a></li></ul></aside></section></section></div>
+<li>Defined in <a href="https://github.com/wwwtyro/candygraph/blob/f289c4d/src/primitives/triangles.ts#L8">primitives/triangles.ts:8</a></li></ul></aside></section></section></div>
 <div class="col-4 col-menu menu-sticky-wrap menu-highlight">
 <div class="tsd-navigation settings">
 <details class="tsd-index-accordion"><summary class="tsd-accordion-summary">
