+++ docs/api/interfaces/VLinesOptions.html
@@ -35,14 +35,14 @@
 <div class="tsd-comment tsd-typography"><p>The color of the lines. If this value is a single Vector4, it will apply to all the lines. Default [0, 0, 0, 1]</p>
 </div><aside class="tsd-sources">
 <ul>
-<li>Defined in <a href="https://github.com/wwwtyro/candygraph/blob/0dd03ce/src/primitives/vlines.ts#L10">primitives/vlines.ts:10</a></li></ul></aside></section>
+<li>Defined in <a href="https://github.com/wwwtyro/candygraph/blob/f289c4d/src/primitives/vlines.ts#L10">primitives/vlines.ts:10</a></li></ul></aside></section>
 <section class="tsd-panel tsd-member tsd-kind-property tsd-parent-kind-interface"><a id="widths" class="tsd-anchor"></a>
 <h3 class="tsd-anchor-link"><code class="tsd-tag ts-flagOptional">Optional</code> <span>widths</span><a href="#widths" aria-label="Permalink" class="tsd-anchor-icon"><svg class="icon icon-tabler icon-tabler-link" viewBox="0 0 24 24" stroke-width="2" stroke="currentColor" fill="none" stroke-linecap="round" stroke-linejoin="round"><use href="#icon-anchor-a"></use><use href="#icon-anchor-b"></use><use href="#icon-anchor-c"></use></svg></a></h3>
 <div class="tsd-signature">widths<span class="tsd-signature-symbol">?:</span> <span class="tsd-signature-type">number</span><span class="tsd-signature-symbol"> | </span><a href="../types/NumberArray.html" class="tsd-signature-type" data-tsd-kind="Type alias">NumberArray</a><span class="tsd-signature-symbol"> | </span><a href="../classes/Dataset.html" class="tsd-signature-type" data-tsd-kind="Class">Dataset</a></div>
 <div class="tsd-comment tsd-typography"><p>The width of the lines in pixels. If this value is a single number, it will apply to all the lines. Default 1.</p>
 </div><aside class="tsd-sources">
 <ul>
-<li>Defined in <a href="https://github.com/wwwtyro/candygraph/blob/0dd03ce/src/primitives/vlines.ts#L8">primitives/vlines.ts:8</a></li></ul></aside></section></section></div>
+<li>Defined in <a href="https://github.com/wwwtyro/candygraph/blob/f289c4d/src/primitives/vlines.ts#L8">primitives/vlines.ts:8</a></li></ul></aside></section></section></div>
 <div class="col-4 col-menu menu-sticky-wrap menu-highlight">
 <div class="tsd-navigation settings">
 <details class="tsd-index-accordion"><summary class="tsd-accordion-summary">
