+++ docs/api/types/Vector2.html
@@ -15,7 +15,7 @@
 <h1>Type alias Vector2</h1></div>
 <div class="tsd-signature">Vector2<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">number</span><span class="tsd-signature-symbol">[]</span><span class="tsd-signature-symbol"> | </span><span class="tsd-signature-type">Float32Array</span></div><aside class="tsd-sources">
 <ul>
-<li>Defined in <a href="https://github.com/wwwtyro/candygraph/blob/0dd03ce/src/common.ts#L73">common.ts:73</a></li></ul></aside></div>
+<li>Defined in <a href="https://github.com/wwwtyro/candygraph/blob/f289c4d/src/common.ts#L73">common.ts:73</a></li></ul></aside></div>
 <div class="col-4 col-menu menu-sticky-wrap menu-highlight">
 <div class="tsd-navigation settings">
 <details class="tsd-index-accordion"><summary class="tsd-accordion-summary">
